requirements.txt
fastapi
uvicorn
requests
aiogram==3.0.0
tortoise-orm